+++ miniprogram/utils/constants.js
@@ -1,6 +1,7 @@
 ﻿module.exports = {
   NORMAL: "normal",
   FAT_LOSS: "fat_loss",
+  SELF_PICK: "self_pick",
   MEAL_TYPE_LABEL: {
     breakfast: "早餐",
     lunch: "中餐",
@@ -8,7 +9,8 @@
   },
   CATEGORY_LABEL: {
     normal: "普通套餐",
-    fat_loss: "减脂套餐"
+    fat_loss: "减脂套餐",
+    self_pick: "自选菜"
   },
   ROLE_LABEL: {
     officer: "民警",